refactor get tasks to a function
